Moves sympmat and is_symplectic into the symplectic module
